Move the tests for generate propety for lyric into "Lyrics" namespace.
