+++ osu.Game.Rulesets.Karaoke.Tests/Editor/Generator/Lyrics/BaseDetectorTest.cs
@@ -6,7 +6,7 @@ using NUnit.Framework;
 using osu.Game.Rulesets.Karaoke.Edit.Generator.Lyrics.Types;
 using osu.Game.Rulesets.Karaoke.Objects;
 
-namespace osu.Game.Rulesets.Karaoke.Tests.Editor.Generator
+namespace osu.Game.Rulesets.Karaoke.Tests.Editor.Generator.Lyrics
 {
     public abstract class BaseDetectorTest<TDetector, TObject, TConfig>
         where TDetector : class, ILyricPropertyDetector<TObject> where TConfig : new()
