+++ osu.Game.Rulesets.Karaoke.Tests/Editor/Generator/Lyrics/BaseGeneratorTest.cs
@@ -6,7 +6,7 @@ using NUnit.Framework;
 using osu.Game.Rulesets.Karaoke.Edit.Generator.Lyrics.Types;
 using osu.Game.Rulesets.Karaoke.Objects;
 
-namespace osu.Game.Rulesets.Karaoke.Tests.Editor.Generator
+namespace osu.Game.Rulesets.Karaoke.Tests.Editor.Generator.Lyrics
 {
     public abstract class BaseGeneratorTest<TGenerator, TObject, TConfig>
         where TGenerator : class, ILyricPropertyGenerator<TObject> where TConfig : new()
